move measurement function into own file
