+++ software/testing/sweep_DAC.py
@@ -1,14 +1,11 @@
 # -*- coding: utf-8 -*-
 
 import sys
-sys.path.append("../pyUSBtin")
-from pyusbtin.usbtin import USBtin
-from pyusbtin.canmessage import CANMessage
-
 sys.path.append("../")
 from OpenFlowMeter import OpenFlowMeter
 from CANsetup import CANsetup
 from PlotData import plotVoltageVsDAC, plotCurrentVsDAC, plotResistanceVsDAC
+from Measurements import DACsweep
 
 import time
 
@@ -52,50 +49,6 @@ def main():
 
     plotCurrentVsDAC(dac[0], current[0])
     plotResistanceVsDAC(dac[0], voltage[0], current[0])
-
-
-def DACsweep(ofm, channel = [0, 1], steps = 128):
-    # initialise dictionaries of arrays to store data
-    dac = {}
-    voltage = {}
-    current = {}
-
-    for chan in channel:
-        voltage[chan] = np.array([])
-        current[chan] = np.array([])
-        dac[chan] = np.array([])
-
-    for DAC in np.linspace(start=0, stop=1023, num=steps, endpoint=True).astype(int):
-        print("%d "%(DAC), end="", flush=True)
-        # set DAC
-        DACset = [0, 0]
-        for chan in channel:
-            dac[chan]= np.append(dac[chan], DAC)
-            DACset[chan] = DAC
-
-        if ofm:
-            ofm.setDAC(DACset[0], DACset[1])
-
-        # wait some seconds to settle everything
-        time.sleep(2)
-
-        # wait until we have a new message
-        if ofm:
-            while not ofm.hasNewMessage:
-                time.sleep(0.5)
-
-        if ofm:
-            for chan in channel:
-                current[chan] = np.append(current[chan], ofm.current(chan))
-                voltage[chan] = np.append(voltage[chan], ofm.voltage(chan))
-
-    # switch off current
-    if ofm:
-        ofm.setDAC(0, 0)
-
-    print("\n...done")
-
-    return dac, voltage, current
 
 if __name__ == "__main__":
     main()
